requirements.txt
flask
flask-cors
pillow
opencv-python-headless
numpy
torch
transformers
